moved over database setup
